+++ apps/sim/app/api/chat/[identifier]/route.ts
@@ -123,11 +123,13 @@ export async function POST(
       const { SSE_HEADERS } = await import('@/lib/utils')
       const { createFilteredResult } = await import('@/app/api/workflows/[id]/execute/route')
 
+      // Generate executionId early so it can be used for file uploads and workflow execution
+      const executionId = crypto.randomUUID()
+
       const workflowInput: any = { input, conversationId }
       if (files && Array.isArray(files) && files.length > 0) {
         logger.debug(`[${requestId}] Processing ${files.length} attached files`)
 
-        const executionId = crypto.randomUUID()
         const executionContext = {
           workspaceId: deployment.userId,
           workflowId: deployment.workflowId,
@@ -153,6 +155,7 @@ export async function POST(
           workflowTriggerType: 'chat',
         },
         createFilteredResult,
+        executionId,
       })
 
       const streamResponse = new NextResponse(stream, {
